Removed method getPrintableString from Person and added it to TextUi
